[screenwavemedia] remove; convert TeamFourStarIE and NormalbootsIE to JWPlatform extractors
